adds audio limit
