+++ client/js/_author/components/assessments/question_types/audio_upload.jsx
@@ -1,46 +1,13 @@
 import React      from 'react';
 import _          from 'lodash';
+import AudioLimit from './question_common/audio_limit';
 import genusTypes from '../../../../constants/genus_types';
 import Feedback   from './question_common/single_feedback';
 
-const audioLimit = (props) => {
-  return null;
-};
-
 export default class AudioUpload extends React.Component {
-  static MAX_TIME = 240;
   static propTypes = {
     updateItem: React.PropTypes.func.isRequired,
     item: React.PropTypes.object
-  }
-
-  static getAudioLimit(item) {
-    const original = _.get(item, ['question', 'timeValue'], {
-      hours: '00',
-      minutes: '00',
-      seconds: '100'
-    });
-    const time = _.mapValues(original, t => parseInt(t, 10));
-    const seconds = (time.hours * 3600) + (time.minutes * 60) + time.seconds;
-    return seconds;
-  }
-
-  static rangeWarning() {
-    return (
-      <span className="au-c-inline-error">Please enter a positive number under 240</span>
-    );
-  }
-
-  constructor(props) {
-    super(props);
-    const timeLimit = AudioUpload.getAudioLimit(props.item)
-      || AudioUpload.MAX_TIME;
-    const displayWarning = timeLimit > AudioUpload.MAX_TIME;
-
-    this.state = {
-      timeLimit: _.toString(timeLimit),
-      displayWarning,
-    };
   }
 
   handleBlur(e) {
@@ -56,48 +23,15 @@ export default class AudioUpload extends React.Component {
     });
   }
 
-  handleTimeLimitUpdate(e) {
-    let timeLimit = e.target.value;
-    const timeVal = parseInt(timeLimit, 10);
-    let displayWarning = false;
-    if (timeVal > AudioUpload.MAX_TIME) {
-      timeLimit = _.toString(AudioUpload.MAX_TIME);
-      displayWarning = true;
-    } else if (_.isNaN(timeVal) || timeVal < 0) {
-      displayWarning = true;
-    }
-
-    this.setState({
-      timeLimit,
-      displayWarning
-    });
-  }
-
   render() {
     return (
       <div>
         <div className="au-c-question__answers au-o-row" role="radiogroup">
           <div className="au-c-file-upload__audio-settings is-active">
-            <span>Audio record limit</span>
-            <div className="au-c-input au-c-input--inline">
-              <label htmlFor="audio-limit" />
-              <div className="au-c-input__contain">
-                <input
-                  className="au-c-text-input au-c-text-input--smaller"
-                  id="audio-limit"
-                  type="text"
-                  maxLength="3"
-                  value={this.state.timeLimit}
-                  onChange={e => this.handleTimeLimitUpdate(e)}
-                  onBlur={e => this.handleBlur(e)}
-                />
-                <div
-                  className={`au-c-input__bottom ${this.state.displayWarning ? 'has-error' : ''}`}
-                />
-              </div>
-            </div>
-            <span>seconds. (240 maximum)</span>
-            { this.state.displayWarning ? AudioUpload.rangeWarning() : null }
+            <AudioLimit
+              handleBlur={e => this.handleBlur(e)}
+              item={this.props.item}
+            />
           </div>
         </div>
         <div className="au-c-question__feedback">
